@@ -406,7 +406,7 @@ public class PageBuilder : IPageBuilder<Page>, IPageBuilder
     }
 
     /// <summary>
-    /// Adds an <see cref="Page"/> field with the provided name and value.
+    /// Adds a <see cref="Page"/> field with the provided name and value.
     /// </summary>
     /// <param name="name">The title of the field.</param>
     /// <param name="value">The value of the field.</param>
@@ -419,7 +419,7 @@ public class PageBuilder : IPageBuilder<Page>, IPageBuilder
     }
 
     /// <summary>
-    /// Adds a field with the provided <see cref="PageBuilder"/> to an <see cref="Page"/>.
+    /// Adds a field with the provided <see cref="PageBuilder"/> to a <see cref="Page"/>.
     /// </summary>
     /// <param name="field">The field builder class containing the field properties.</param>
     /// <returns>The current builder.</returns>
@@ -430,7 +430,7 @@ public class PageBuilder : IPageBuilder<Page>, IPageBuilder
     }
 
     /// <summary>
-    /// Adds an <see cref="Page"/> field with the provided properties.
+    /// Adds a <see cref="Page"/> field with the provided properties.
     /// </summary>
     /// <param name="action">The delegate containing the field properties.</param>
     /// <returns>The current builder.</returns>
@@ -544,36 +544,37 @@ public class PageBuilder : IPageBuilder<Page>, IPageBuilder
     /// <inheritdoc/>
     IPage IPageBuilder<IPage>.Build() => Build();
 
-    internal PageBuilder WithPaginatorFooter(PaginatorFooter footer, int page, int totalPages, ICollection<IUser>? users)
+    internal PageBuilder WithPaginatorFooter(PaginatorFooter footer, int currentPageIndex, int maxPageIndex, ICollection<IUser>? users)
     {
         if (footer == PaginatorFooter.None)
-        {
             return this;
-        }
 
         Footer = new EmbedFooterBuilder();
+
         if (footer.HasFlag(PaginatorFooter.Users))
         {
             if (users is null || users.Count == 0)
             {
-                Footer.Text += "Interactors: Everyone\n";
+                Footer.Text += "Interactors: Everyone";
             }
             else if (users.Count == 1)
             {
-                var user = users.First();
+                var user = users.Single();
 
-                Footer.IconUrl = user.GetAvatarUrl() ?? user.GetDefaultAvatarUrl();
-                Footer.Text += $"Interactor: {user}\n";
+                Footer.Text += $"Interactor: {user}";
+                Footer.IconUrl = user.GetDisplayAvatarUrl();
             }
             else
             {
                 Footer.Text += $"Interactors: {string.Join(", ", users)}";
             }
+
+            Footer.Text += '\n';
         }
 
         if (footer.HasFlag(PaginatorFooter.PageNumber))
         {
-            Footer.Text += $"Страница: {page + 1} из {totalPages + 1}";
+            Footer.Text += $"Страница: {currentPageIndex + 1} из {maxPageIndex + 1}";
         }
 
         return this;
